-added reindex catalog button (temporarily) -fixed bugs with reindex catalog: --fixed bug with duplicating database link --fixed bug with NaN sized database --fixed bug when updated database view is not updated -fixed bug when user choose wrong (or broken) image -added path and model size to edit properties modal -added menu links to wiki and release notes -added new formats: 3b, spp -added icons -updated styles -updated localization -updated type definitions

+++ src/views/Modals/Edit/EditPropertiesModal.vue
@@ -4,71 +4,91 @@
       <h2>{{ $t('context.model.properties') }}</h2>
     </header>
     <div class="modal_content">
-      <img
-        class="modal_image"
-        :src="forceReloadImage(properties.image)"
-      >
-      <div class="input-group">
-        <label>{{ $t('modals.properties.title') }}</label>
-        <input
-          v-model="properties.name"
-          type="text"
-          class="input"
-          required
+      <model-image
+        :src="properties.image"
+        :extension="properties.extension"
+      />
+      <ValidationObserver ref="form">
+        <ValidationProvider
+          v-slot="{ classes }"
+          name="modelTitle"
+          class="input-group"
+          rules="required"
+          immediate
         >
-      </div>
-      <div class="input-group">
-        <label>{{ $t('modals.properties.category') }}</label>
-        <div class="input--select">
-          <select
-            v-model="properties.category"
+          <label>{{ $t('modals.properties.title') }}</label>
+          <input
+            v-model="properties.name"
+            type="text"
             class="input"
+            :class="classes"
+            :placeholder="$t('modals.properties.title')"
           >
-            <option :value="null">
-              Без категории
-            </option>
-            <option
-              v-for="category in categories"
-              :key="category.id"
-              :value="category.id"
+        </ValidationProvider>
+        <div class="input-group">
+          <label>{{ $t('modals.properties.category') }}</label>
+          <div class="input--select">
+            <select
+              v-model="properties.category"
+              class="input"
             >
-              {{ rewriteOptionName(category) }}
-            </option>
-          </select>
-          <vue-icon
-            icon="caret-forward"
-            raster
-          />
+              <option :value="null">
+                {{ $t('views.catalog.categories.noCategory') }}
+              </option>
+              <option
+                v-for="category in categories"
+                :key="category.id"
+                :value="category.id"
+              >
+                {{ rewriteOptionName(category) }}
+              </option>
+            </select>
+            <vue-icon
+              icon="caret-forward"
+              raster
+            />
+          </div>
         </div>
-      </div>
-      <div class="input-group">
-        <label>{{ $t('modals.properties.image') }}</label>
-        <input
-          ref="file"
-          type="file"
-          accept=".jpg, .jpeg, .png, .webp"
-          hidden
-          required
-          @change="handleFileChange"
-        >
-        <button
-          class="input input--file"
-          @click.prevent="handleFileClick"
-        >
-          <span v-if="uploadImage !== ''">{{ uploadImage }}</span>
-          <span
-            v-else
-            class="placeholder"
+        <div class="input-group">
+          <label>{{ $t('modals.properties.image') }}</label>
+          <input
+            ref="file"
+            type="file"
+            accept=".jpg, .jpeg, .png, .webp"
+            hidden
+            required
+            @change="handleFileChange"
           >
-            {{ $t('modals.properties.image') }}
-          </span>
-        </button>
-      </div>
+          <button
+            class="input input--file"
+            @click.prevent="handleFileClick"
+          >
+            <span v-if="uploadImage !== ''">{{ uploadImage }}</span>
+            <span
+              v-else
+              class="placeholder"
+            >
+              {{ $t('modals.properties.image') }}
+            </span>
+          </button>
+        </div>
+        <div class="info">
+          <dd>{{ $t('views.catalog.models.path') }}</dd>
+          <dt>{{ properties.path }}</dt>
+        </div>
+        <div class="info">
+          <dd>{{ $t('views.catalog.models.size') }}</dd>
+          <dt>{{ $formatSize(properties.size) }}</dt>
+        </div>
+        <div class="errors">
+          {{ error }}
+        </div>
+      </ValidationObserver>
     </div>
     <div class="modal_actions">
       <button
         class="button button--primary"
-        @click="updateItem()"
+        @click="handleSubmit"
       >
         {{ $t('common.buttons.save') }}
       </button>
@@ -92,22 +112,37 @@ import sharp from 'sharp'
 import uniqid from 'uniqid'
 import { remote } from 'electron'
 import Integrations from '@/plugins/models-db/integrations/main'
+import ModelImage from '@/components/UI/Image/ModelImage.vue'
+import { ValidationObserver, ValidationProvider } from 'vee-validate'
 
-@Component({})
+@Component({
+  components: {
+    ModelImage,
+    ValidationObserver,
+    ValidationProvider
+  }
+})
 export default class EditPropertiesModal extends Vue {
-  categories: any[] = []
+  error = ''
+  categories: Category[] = []
   uploadImage = ''
   properties: ImageProperties = {
-    category: '',
+    id: 0,
+    category: null,
     extension: '',
     name: '',
     image: '',
     imageChanged: false,
     path: '',
+    size: 0
   }
 
   async mounted(): Promise<void> {
     this.properties = this.$store.state.controls.properties
+
+    if (this.properties.category === '') {
+      this.properties.category = null
+    }
 
     const db = new Integrations.local(this.$route.params.database)
     const categories = await db.fetchCategories(`SELECT * FROM 'categories'`)
@@ -115,7 +150,7 @@ export default class EditPropertiesModal extends Vue {
     this.categories = categories
   }
 
-  rewriteOptionName(category: any): string {
+  rewriteOptionName(category: Category): string {
     return category.parentId !== -1
     ? `${this.getParentName(category.parentId)}${category.name}`
     : category.name
@@ -124,7 +159,7 @@ export default class EditPropertiesModal extends Vue {
   getParentName(parentId: number): string {
     const matches: string[] = []
 
-    this.categories.forEach((item: any) => {
+    this.categories.forEach((item: Category) => {
       if (item.id === parentId) {
         matches.push(item.name)
       }
@@ -148,6 +183,13 @@ export default class EditPropertiesModal extends Vue {
     (this.$refs.file as HTMLInputElement).click()
   }
 
+  async handleSubmit(): Promise<void> {
+    const success = await (this.$refs.form as InstanceType<typeof ValidationObserver>).validate()
+    if (success) {
+      await this.updateItem()
+    }
+  }
+
   async updateItem(): Promise<void> {
     const db = new Integrations.local(this.$route.params.database)
 
@@ -157,14 +199,13 @@ export default class EditPropertiesModal extends Vue {
     const imageName = uniqid('image-') + '.webp'
 
     let imagePath = ''
+    this.error = ''
     if (this.properties.imageChanged === true && this.properties.image !== '') {
       const image = path.parse(this.properties.image)
       image.ext = '.webp'
-      imagePath = path.normalize(path.join(image.dir, image.name) + image.ext)
+      imagePath = path.join(image.dir, image.name) + image.ext
     } else {
-      imagePath = path.normalize(
-        path.join(remote.app.getPath('userData'), '\\imagecache\\', imageName)
-      )
+      imagePath = path.join(remote.app.getPath('userData'), '\\imagecache\\', imageName)
     }
 
     models[0].name = this.properties.name
@@ -175,7 +216,9 @@ export default class EditPropertiesModal extends Vue {
     // Create thumbnails and save in imagecache folder
     if (this.properties.imageChanged === true) {
       fs.access(imagePath, fs.constants.F_OK, (err) => {
-        fs.unlinkSync(this.properties.image)
+        if(fs.existsSync(this.properties.image)) {
+          fs.unlinkSync(this.properties.image)
+        }
         sharp(this.uploadImage)
           .resize(1024, 1024)
           .webp({
@@ -191,6 +234,17 @@ export default class EditPropertiesModal extends Vue {
             const result = await db.fetchItemsFromDatabase()
             this.$store.commit('setLoadedData', result)
             this.$emit('close')
+          })
+          .catch(async (err: Error) => {
+            this.error = err.name + ': ' + err.message
+            this.properties.image = ''
+            this.properties.imageChanged = false
+
+            models[0].image = ''
+
+            this.$updateItemInDatabase(this.$route.params.database, models[0])
+            const result = await db.fetchItemsFromDatabase()
+            this.$store.commit('setLoadedData', result)
           })
       })
     } else {
